+++ CHANGELOG.rst
@@ -17,3 +17,4 @@ first official Alpha release.
  * Added jquery.cicle in media. jQuery plugin to handle rotating pictures.
  * Added plugin's url for handling dynamic plugins not connected to a model.
  * Added Contact form plugin.
+ * ADDED NEW MODEL AdditionalPageData for injecting ad-hoc data into page head or javascript in page body